requirements.txt
# Core
pandas>=2.0
numpy>=1.24
scipy>=1.10

# ML
scikit-learn>=1.3
imbalanced-learn>=0.11
xgboost>=1.7

# Visualization (optional but used)
matplotlib>=3.7
seaborn>=0.12

# Date parsing
python-dateutil>=2.8

# Database
psycopg2-binary>=2.9

# AI
google-generativeai>=0.3
jsonschema>=4.21
pyyaml>=6.0

# UI
streamlit>=1.30
